split tpc builder to add laser builder
